[IOTDB-3436] Add IT for aggregation, groupby and align by device query in MPP (#6207)
