+++ integration-test/src/test/java/org/apache/iotdb/db/it/aligned/IoTDBAggregationWithoutValueFilterWithDeletionIT.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -16,21 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.integration.aligned;
+package org.apache.iotdb.db.it.aligned;
 
-import org.apache.iotdb.db.conf.IoTDBDescriptor;
-import org.apache.iotdb.db.utils.EnvironmentUtils;
-import org.apache.iotdb.itbase.category.LocalStandaloneTest;
-import org.apache.iotdb.jdbc.Config;
+import org.apache.iotdb.it.env.ConfigFactory;
+import org.apache.iotdb.it.env.EnvFactory;
+import org.apache.iotdb.itbase.category.ClusterIT;
+import org.apache.iotdb.itbase.category.LocalStandaloneIT;
 
 import org.junit.AfterClass;
-import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
@@ -42,8 +40,8 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
-@Category({LocalStandaloneTest.class})
-public class IoTDBAggregationWithoutValueFilterIT {
+@Category({LocalStandaloneIT.class, ClusterIT.class})
+public class IoTDBAggregationWithoutValueFilterWithDeletionIT {
 
   private static final double DELTA = 1e-6;
   protected static boolean enableSeqSpaceCompaction;
@@ -52,57 +50,54 @@ public class IoTDBAggregationWithoutValueFilterIT {
 
   @BeforeClass
   public static void setUp() throws Exception {
-    EnvironmentUtils.envSetUp();
-    // TODO When the aligned time series support compaction, we need to set compaction to true
-    enableSeqSpaceCompaction =
-        IoTDBDescriptor.getInstance().getConfig().isEnableSeqSpaceCompaction();
-    enableUnseqSpaceCompaction =
-        IoTDBDescriptor.getInstance().getConfig().isEnableUnseqSpaceCompaction();
-    enableCrossSpaceCompaction =
-        IoTDBDescriptor.getInstance().getConfig().isEnableCrossSpaceCompaction();
-    IoTDBDescriptor.getInstance().getConfig().setEnableSeqSpaceCompaction(false);
-    IoTDBDescriptor.getInstance().getConfig().setEnableUnseqSpaceCompaction(false);
-    IoTDBDescriptor.getInstance().getConfig().setEnableCrossSpaceCompaction(false);
+    enableSeqSpaceCompaction = ConfigFactory.getConfig().isEnableSeqSpaceCompaction();
+    enableUnseqSpaceCompaction = ConfigFactory.getConfig().isEnableUnseqSpaceCompaction();
+    enableCrossSpaceCompaction = ConfigFactory.getConfig().isEnableCrossSpaceCompaction();
+    ConfigFactory.getConfig().setEnableSeqSpaceCompaction(false);
+    ConfigFactory.getConfig().setEnableUnseqSpaceCompaction(false);
+    ConfigFactory.getConfig().setEnableCrossSpaceCompaction(false);
+    EnvFactory.getEnv().initBeforeClass();
     AlignedWriteUtil.insertData();
+
+    try (Connection connection = EnvFactory.getEnv().getConnection();
+        Statement statement = connection.createStatement()) {
+      // TODO replace it while delete timeseries is supported in cluster mode
+      //      statement.execute("delete timeseries root.sg1.d1.s2");
+      statement.execute("delete from root.sg1.d1.s2 where time <= 40");
+      statement.execute("delete from root.sg1.d1.s1 where time <= 21");
+    } catch (Exception e) {
+      fail(e.getMessage());
+      e.printStackTrace();
+    }
   }
 
   @AfterClass
   public static void tearDown() throws Exception {
-    IoTDBDescriptor.getInstance().getConfig().setEnableSeqSpaceCompaction(enableSeqSpaceCompaction);
-    IoTDBDescriptor.getInstance()
-        .getConfig()
-        .setEnableUnseqSpaceCompaction(enableUnseqSpaceCompaction);
-    IoTDBDescriptor.getInstance()
-        .getConfig()
-        .setEnableCrossSpaceCompaction(enableCrossSpaceCompaction);
-    EnvironmentUtils.cleanEnv();
+    EnvFactory.getEnv().cleanAfterClass();
+    ConfigFactory.getConfig().setEnableSeqSpaceCompaction(enableSeqSpaceCompaction);
+    ConfigFactory.getConfig().setEnableUnseqSpaceCompaction(enableUnseqSpaceCompaction);
+    ConfigFactory.getConfig().setEnableCrossSpaceCompaction(enableCrossSpaceCompaction);
   }
 
   @Test
-  public void countAllAlignedWithoutTimeFilterTest() throws ClassNotFoundException {
-    String[] retArray = new String[] {"20", "29", "28", "19", "20"};
+  public void countAllAlignedWithoutTimeFilterTest() {
+    String[] retArray = new String[] {"1", "28", "19", "20"};
     String[] columnNames = {
       "count(root.sg1.d1.s1)",
-      "count(root.sg1.d1.s2)",
       "count(root.sg1.d1.s3)",
       "count(root.sg1.d1.s4)",
       "count(root.sg1.d1.s5)"
     };
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("select count(*) from root.sg1.d1");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("select count(*) from root.sg1.d1")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>(); // used to adjust result sequence
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           String[] ans = new String[columnNames.length];
@@ -124,11 +119,10 @@ public class IoTDBAggregationWithoutValueFilterIT {
   }
 
   @Test
-  public void countAllAlignedAndNonAlignedWithoutTimeFilterTest() throws ClassNotFoundException {
-    String[] retArray = new String[] {"20", "29", "28", "19", "20", "19", "29", "28", "18", "19"};
+  public void countAllAlignedAndNonAlignedWithoutTimeFilterTest() {
+    String[] retArray = new String[] {"1", "28", "19", "20", "19", "29", "28", "18", "19"};
     String[] columnNames = {
       "count(root.sg1.d1.s1)",
-      "count(root.sg1.d1.s2)",
       "count(root.sg1.d1.s3)",
       "count(root.sg1.d1.s4)",
       "count(root.sg1.d1.s5)",
@@ -138,21 +132,16 @@ public class IoTDBAggregationWithoutValueFilterIT {
       "count(root.sg1.d2.s4)",
       "count(root.sg1.d2.s5)"
     };
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("select count(*) from root.sg1.*");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("select count(*) from root.sg1.*")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           String[] ans = new String[columnNames.length];
@@ -174,31 +163,26 @@ public class IoTDBAggregationWithoutValueFilterIT {
   }
 
   @Test
-  public void countAllAlignedWithTimeFilterTest() throws ClassNotFoundException {
-    String[] retArray = new String[] {"12", "15", "22", "13", "6"};
+  public void countAllAlignedWithTimeFilterTest() {
+    String[] retArray = new String[] {"1", "22", "13", "6"};
     String[] columnNames = {
       "count(root.sg1.d1.s1)",
-      "count(root.sg1.d1.s2)",
       "count(root.sg1.d1.s3)",
       "count(root.sg1.d1.s4)",
       "count(root.sg1.d1.s5)"
     };
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet =
-          statement.execute("select count(*) from root.sg1.d1 where time >= 9 and time <= 33");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(*) from root.sg1.d1 where time >= 9 and time <= 33")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           String[] ans = new String[columnNames.length];
@@ -221,39 +205,28 @@ public class IoTDBAggregationWithoutValueFilterIT {
 
   /** aggregate multi columns of aligned timeseries in one SQL */
   @Test
-  public void aggregateSomeAlignedWithoutTimeFilterTest() throws ClassNotFoundException {
-    double[] retArray =
-        new double[] {
-          20, 29, 28, 390184, 130549, 390417, 19509.2, 4501.689655172413, 13943.464285714286
-        };
+  public void aggregateSomeAlignedWithoutTimeFilterTest() {
+    double[] retArray = new double[] {1, 28, 230000, 390417, 230000, 13943.464285714286};
     String[] columnNames = {
       "count(root.sg1.d1.s1)",
-      "count(root.sg1.d1.s2)",
       "count(root.sg1.d1.s3)",
       "sum(root.sg1.d1.s1)",
-      "sum(root.sg1.d1.s2)",
       "sum(root.sg1.d1.s3)",
       "avg(root.sg1.d1.s1)",
-      "avg(root.sg1.d1.s2)",
       "avg(root.sg1.d1.s3)",
     };
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet =
-          statement.execute(
-              "select count(s1),count(s2),count(s3),sum(s1),sum(s2),sum(s3),avg(s1),avg(s2),avg(s3) from root.sg1.d1");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(s1),count (s3),sum(s1),sum(s3),avg(s1),avg(s3) from root.sg1.d1")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           double[] ans = new double[columnNames.length];
@@ -276,39 +249,28 @@ public class IoTDBAggregationWithoutValueFilterIT {
 
   /** aggregate multi columns of aligned timeseries in one SQL */
   @Test
-  public void aggregateSomeAlignedWithTimeFilterTest() throws ClassNotFoundException {
-    double[] retArray =
-        new double[] {
-          6, 9, 15, 230090, 220, 230322, 38348.333333333336, 24.444444444444443, 15354.8
-        };
+  public void aggregateSomeAlignedWithTimeFilterTest() {
+    double[] retArray = new double[] {1, 15, 230000, 230322, 230000, 15354.8};
     String[] columnNames = {
       "count(root.sg1.d1.s1)",
-      "count(root.sg1.d1.s2)",
       "count(root.sg1.d1.s3)",
       "sum(root.sg1.d1.s1)",
-      "sum(root.sg1.d1.s2)",
       "sum(root.sg1.d1.s3)",
       "avg(root.sg1.d1.s1)",
-      "avg(root.sg1.d1.s2)",
       "avg(root.sg1.d1.s3)",
     };
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet =
-          statement.execute(
-              "select count(s1),count (s2),count (s3),sum(s1),sum(s2),sum(s3),avg(s1),avg(s2),avg(s3) from root.sg1.d1 where time>=16 and time<=34");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(s1),count (s3),sum(s1),sum(s3),avg(s1),avg(s3) from root.sg1.d1 where time>=16 and time<=34")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           double[] ans = new double[columnNames.length];
@@ -330,25 +292,20 @@ public class IoTDBAggregationWithoutValueFilterIT {
   }
 
   @Test
-  public void countSingleAlignedWithTimeFilterTest() throws ClassNotFoundException {
-    String[] retArray = new String[] {"9"};
-    String[] columnNames = {"count(root.sg1.d1.s2)"};
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void countSingleAlignedWithTimeFilterTest() {
+    String[] retArray = new String[] {"0"};
+    String[] columnNames = {"count(root.sg1.d1.s1)"};
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet =
-          statement.execute("select count(s2) from root.sg1.d1 where time>=16 and time<=34");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery("select count(s1) from root.sg1.d1 where time>=16 and time<=20")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           StringBuilder builder = new StringBuilder();
@@ -372,25 +329,20 @@ public class IoTDBAggregationWithoutValueFilterIT {
   }
 
   @Test
-  public void sumSingleAlignedWithTimeFilterTest() throws ClassNotFoundException {
-    String[] retArray = new String[] {"230322.0"};
-    String[] columnNames = {"sum(root.sg1.d1.s3)"};
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void sumSingleAlignedWithTimeFilterTest() {
+    String[] retArray = new String[] {"null"};
+    String[] columnNames = {"sum(root.sg1.d1.s1)"};
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet =
-          statement.execute("select sum(s3) from root.sg1.d1 where time>=16 and time<=34");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery("select sum(s1) from root.sg1.d1 where time>=16 and time<=20")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
           StringBuilder builder = new StringBuilder();
@@ -414,36 +366,32 @@ public class IoTDBAggregationWithoutValueFilterIT {
   }
 
   @Test
-  public void avgSingleAlignedWithTimeFilterTest() throws ClassNotFoundException {
-    double[][] retArray = {{24.444444444444443}};
-    String[] columnNames = {"avg(root.sg1.d1.s2)"};
-    Class.forName(Config.JDBC_DRIVER_NAME);
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void avgSingleAlignedWithTimeFilterTest() {
+    String[] retArray = new String[] {"null"};
+    String[] columnNames = {"avg(root.sg1.d1.s1)"};
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet =
-          statement.execute("select avg(s2) from root.sg1.d1 where time>=16 and time<=34");
-      Assert.assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery("select avg(s1) from root.sg1.d1 where time>=16 and time<=20")) {
         ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
         Map<String, Integer> map = new HashMap<>();
         for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
           map.put(resultSetMetaData.getColumnName(i), i);
         }
-        assertEquals(columnNames.length, resultSetMetaData.getColumnCount());
+
         int cnt = 0;
         while (resultSet.next()) {
-          double[] ans = new double[columnNames.length];
           StringBuilder builder = new StringBuilder();
           // No need to add time column for aggregation query
-          for (int i = 0; i < columnNames.length; i++) {
-            String columnName = columnNames[i];
+          for (String columnName : columnNames) {
             int index = map.get(columnName);
-            ans[i] = Double.parseDouble(resultSet.getString(index));
+            if (builder.length() != 0) {
+              builder.append(",");
+            }
+            builder.append(resultSet.getString(index));
           }
-          assertArrayEquals(retArray[cnt], ans, DELTA);
+          assertEquals(retArray[cnt], builder.toString());
           cnt++;
         }
         assertEquals(1, cnt);
