+++ integration-test/src/test/java/org/apache/iotdb/db/it/aligned/IoTDBGroupByQueryWithoutValueFilterWithDeletionIT.java
@@ -16,13 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.integration.aligned;
+package org.apache.iotdb.db.it.aligned;
 
-import org.apache.iotdb.db.conf.IoTDBDescriptor;
-import org.apache.iotdb.integration.env.ConfigFactory;
-import org.apache.iotdb.integration.env.EnvFactory;
-import org.apache.iotdb.itbase.category.LocalStandaloneTest;
-import org.apache.iotdb.jdbc.Config;
+import org.apache.iotdb.it.env.ConfigFactory;
+import org.apache.iotdb.it.env.EnvFactory;
+import org.apache.iotdb.itbase.category.ClusterIT;
+import org.apache.iotdb.itbase.category.LocalStandaloneIT;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
@@ -43,7 +42,7 @@ import static org.apache.iotdb.db.constant.TestConstant.maxValue;
 import static org.apache.iotdb.db.constant.TestConstant.minTime;
 import static org.apache.iotdb.db.constant.TestConstant.minValue;
 
-@Category({LocalStandaloneTest.class})
+@Category({LocalStandaloneIT.class, ClusterIT.class})
 public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
 
   protected static boolean enableSeqSpaceCompaction;
@@ -55,23 +54,21 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
 
   @BeforeClass
   public static void setUp() throws Exception {
-    EnvFactory.getEnv().initBeforeClass();
-    enableSeqSpaceCompaction =
-        IoTDBDescriptor.getInstance().getConfig().isEnableSeqSpaceCompaction();
-    enableUnseqSpaceCompaction =
-        IoTDBDescriptor.getInstance().getConfig().isEnableUnseqSpaceCompaction();
-    enableCrossSpaceCompaction =
-        IoTDBDescriptor.getInstance().getConfig().isEnableCrossSpaceCompaction();
-    prevPartitionInterval = IoTDBDescriptor.getInstance().getConfig().getPartitionInterval();
+    enableSeqSpaceCompaction = ConfigFactory.getConfig().isEnableSeqSpaceCompaction();
+    enableUnseqSpaceCompaction = ConfigFactory.getConfig().isEnableUnseqSpaceCompaction();
+    enableCrossSpaceCompaction = ConfigFactory.getConfig().isEnableCrossSpaceCompaction();
     ConfigFactory.getConfig().setEnableSeqSpaceCompaction(false);
     ConfigFactory.getConfig().setEnableUnseqSpaceCompaction(false);
     ConfigFactory.getConfig().setEnableCrossSpaceCompaction(false);
+    EnvFactory.getEnv().initBeforeClass();
     AlignedWriteUtil.insertData();
-    Class.forName(Config.JDBC_DRIVER_NAME);
+
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("delete from root.sg1.d1.s1 where time <= 15");
-      statement.execute("delete timeseries root.sg1.d1.s2");
+      // TODO replace it while delete timeseries is supported in cluster mode
+      //      statement.execute("delete timeseries root.sg1.d1.s2");
+      statement.execute("delete from root.sg1.d1.s2 where time <= 40");
       statement.execute("delete from root.sg1.d1.s3 where time > 25");
     } catch (Exception e) {
       e.printStackTrace();
@@ -92,14 +89,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
     String[] retArray = new String[] {"1,0,8.0", "11,5,13014.2", "21,1,46018.4", "31,0,null"};
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
-                  + "where time > 5 GROUP BY ([1, 41), 10ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
+                  + "where time > 5 GROUP BY ([1, 41), 10ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -114,13 +109,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
-                  + " where time > 5 GROUP BY ([1, 41), 10ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 GROUP BY ([1, 41), 10ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -146,14 +138,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
-                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 5ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
+                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 5ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -168,13 +158,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
-                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 5ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 5ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -199,14 +186,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
-                  + "where time > 5 GROUP BY ([1, 41), 4ms, 6ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
+                  + "where time > 5 GROUP BY ([1, 41), 4ms, 6ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -221,13 +206,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select count(s1), sum(s2), avg(s3) from root.sg1.d1 "
-                  + " where time > 5 GROUP BY ([1, 41), 4ms, 6ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 GROUP BY ([1, 41), 4ms, 6ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -252,14 +234,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
-                  + "where time > 5 GROUP BY ([1, 41), 10ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
+                  + "where time > 5 GROUP BY ([1, 41), 10ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -276,13 +256,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
-                  + " where time > 5 GROUP BY ([1, 41), 10ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 GROUP BY ([1, 41), 10ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -316,14 +293,11 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
-                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 5ms)");
-      Assert.assertTrue(hasResultSet);
-
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
+                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 5ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -340,13 +314,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
-                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 5ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 5ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -379,14 +350,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
-                  + "where time > 5 GROUP BY ([1, 41), 4ms, 6ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
+                  + "where time > 5 GROUP BY ([1, 41), 4ms, 6ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -403,13 +372,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select max_value(s3), min_value(s1), max_time(s2), min_time(s3) from root.sg1.d1 "
-                  + " where time > 5 GROUP BY ([1, 41), 4ms, 6ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 GROUP BY ([1, 41), 4ms, 6ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -433,14 +399,11 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
     String[] retArray = new String[] {"1,null,6", "11,20.0,11", "21,230000.0,21", "31,null,null"};
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select last_value(s1), first_value(s3) from root.sg1.d1 "
-                  + "where time > 5 GROUP BY ([1, 41), 10ms)");
-      Assert.assertTrue(hasResultSet);
-
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select last_value(s1), first_value(s3) from root.sg1.d1 "
+                  + "where time > 5 GROUP BY ([1, 41), 10ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -455,13 +418,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select last_value(s1), first_value(s3) from root.sg1.d1 "
-                  + " where time > 5 GROUP BY ([1, 41), 10ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 GROUP BY ([1, 41), 10ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -493,14 +453,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select last_value(s1), first_value(s3) from root.sg1.d1 "
-                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 5ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select last_value(s1), first_value(s3) from root.sg1.d1 "
+                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 5ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -515,13 +473,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select last_value(s1), first_value(s3) from root.sg1.d1 "
-                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 5ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 5ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -552,14 +507,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select last_value(s1), first_value(s3) from root.sg1.d1 "
-                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 4ms, 6ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select last_value(s1), first_value(s3) from root.sg1.d1 "
+                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 4ms, 6ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -574,13 +527,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select last_value(s1), first_value(s3) from root.sg1.d1 "
-                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 4ms, 6ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 4ms, 6ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -608,13 +558,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select count(*), last_value(*) from root.sg1.d1 GROUP BY ([1, 41), 10ms)");
-      Assert.assertTrue(hasResultSet);
-
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(*), last_value(*) from root.sg1.d1 GROUP BY ([1, 41), 10ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -641,13 +588,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select count(*), last_value(*) from root.sg1.d1 "
-                  + "GROUP BY ([1, 41), 10ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + "GROUP BY ([1, 41), 10ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -693,14 +637,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select count(*) from root.sg1.d1 "
-                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select count(*) from root.sg1.d1 "
+                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -719,13 +661,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select count(*) from root.sg1.d1 "
-                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
@@ -763,14 +702,12 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         };
     try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
-      boolean hasResultSet =
-          statement.execute(
-              "select last_value(*) from root.sg1.d1 "
-                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms)");
-      Assert.assertTrue(hasResultSet);
 
       int cnt;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery(
+              "select last_value(*) from root.sg1.d1 "
+                  + "where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms)")) {
         cnt = 0;
         while (resultSet.next()) {
           String ans =
@@ -789,13 +726,10 @@ public class IoTDBGroupByQueryWithoutValueFilterWithDeletionIT {
         Assert.assertEquals(retArray.length, cnt);
       }
 
-      hasResultSet =
-          statement.execute(
+      try (ResultSet resultSet =
+          statement.executeQuery(
               "select last_value(*) from root.sg1.d1 "
-                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms) order by time desc");
-      Assert.assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+                  + " where time > 5 and time < 38 GROUP BY ([1, 41), 3ms, 4ms) order by time desc")) {
         cnt = retArray.length;
         while (resultSet.next()) {
           String ans =
