+++ server/src/main/java/org/apache/iotdb/db/mpp/aggregation/CountAccumulator.java
@@ -61,6 +61,9 @@ public class CountAccumulator implements Accumulator {
 
   @Override
   public void addStatistics(Statistics statistics) {
+    if (statistics == null) {
+      return;
+    }
     countValue += statistics.getCount();
   }
