+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/planner/plan/node/source/AlignedSeriesAggregationScanNode.java
@@ -289,7 +289,7 @@ public class AlignedSeriesAggregationScanNode extends SeriesAggregationSourceNod
   @Override
   public String toString() {
     return String.format(
-        "SeriesAggregationScanNode-%s:[SeriesPath: %s, Descriptor: %s, DataRegion: %s]",
+        "AlignedSeriesAggregationScanNode-%s:[SeriesPath: %s, Descriptor: %s, DataRegion: %s]",
         this.getPlanNodeId(),
         this.getAlignedPath().getFormattedString(),
         this.getAggregationDescriptorList(),
